refactor: move GitLab and Bitrix24 client services to Domain
Move infrastructure client services to their bounded contexts:
- GitLab: GitLabClient(Interface), BranchParser
- GitLab: ConventionalCommitParser (from Services/Sync/)
- Bitrix24: Bitrix24Client(Interface)
