+++ backend/app/Domain/GitLab/Services/GitLabClientInterface.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Services\GitLab;
+namespace App\Domain\GitLab\Services;
 
 interface GitLabClientInterface
 {
