+++ backend/app/Services/Sync/SyncService.php
@@ -12,11 +12,12 @@ use App\Domain\GitLab\Models\Branch;
 use App\Domain\GitLab\Models\Commit;
 use App\Domain\Settings\Models\ProjectMapping;
 use App\Domain\Settings\Models\Setting;
+use App\Domain\GitLab\Services\ConventionalCommitParser;
 use App\Domain\Sync\Models\SyncLog;
 use App\Domain\Bitrix24\Models\Task;
-use App\Services\Bitrix24\Bitrix24ClientInterface;
-use App\Services\GitLab\BranchParser;
-use App\Services\GitLab\GitLabClientInterface;
+use App\Domain\Bitrix24\Services\Bitrix24ClientInterface;
+use App\Domain\GitLab\Services\BranchParser;
+use App\Domain\GitLab\Services\GitLabClientInterface;
 use App\Services\Matching\MatchingEngineInterface;
 use Carbon\CarbonImmutable;
 use Illuminate\Support\Facades\Log;
