+++ backend/tests/Unit/Services/Sync/SyncServiceTest.php
@@ -12,11 +12,11 @@ use App\Domain\GitLab\Models\Commit;
 use App\Domain\Settings\Models\ProjectMapping;
 use App\Domain\Settings\Models\Setting;
 use App\Domain\Bitrix24\Models\Task;
-use App\Services\Bitrix24\Bitrix24ClientInterface;
-use App\Services\GitLab\BranchParser;
-use App\Services\GitLab\GitLabClientInterface;
+use App\Domain\Bitrix24\Services\Bitrix24ClientInterface;
+use App\Domain\GitLab\Services\BranchParser;
+use App\Domain\GitLab\Services\GitLabClientInterface;
 use App\Services\Matching\MatchingEngineInterface;
-use App\Services\Sync\ConventionalCommitParser;
+use App\Domain\GitLab\Services\ConventionalCommitParser;
 use App\Services\Sync\SyncService;
 use Illuminate\Foundation\Testing\RefreshDatabase;
 use Illuminate\Support\Collection;
